+++ Scripts/Buildings/GathererHutBuildingLoader.cs
@@ -56,7 +56,7 @@ public class GathererHutBuildingData : ABuildingData
     public string[] recipes;
 
     [SchemaField(null)]
-    public float? maxStorage;
+    public int? maxStorage;
 
     [SchemaField(null)]
     public float? maxDistance;
